split monolithic streamableresource.py into easier to understand individual files
